requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
config==0.5.1
cssselect==1.2.0
discord.py==2.3.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.3
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
pandas==2.1.0
Pillow==10.0.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydub==0.25.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.12.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.0
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.2
youtube-transcript-api==0.6.1
